fixed routes (mostly) + added descriptions

+++ server/router.ts
@@ -21,7 +21,7 @@ import {TemplateI} from "./interfaces/template";
 let router = express.Router();
 router.use(bodyParser.json());
 
-let upload = multer({ dest: '/tmp/' });
+let upload = multer({dest: '/tmp/'});
 
 export = router;
 
@@ -50,7 +50,7 @@ router.post('/api/latex/convert', (req, res) => {
   let name = req.body.name;
   let latex = req.body.latex;
   if (name && latex) {
-    require('node-latex')(latex, { inputs: Config.DATA_PATH + name + '/' })
+    require('node-latex')(latex, {inputs: Config.DATA_PATH + name + '/'})
       .on('error', err => {
         console.error(err);
         res.status(500).send({error: err.toString()});
@@ -83,30 +83,30 @@ router.post('/api/upload', upload.single('file'), (req, res) => {
 /**
  * CRUD system for managing different latex documents.
  */
-interface File {
+interface mFile {
   name: string,
   text: string,
 }
+
 // Get all documents, returns a list of available templates in the db.
-router.get('/api/latex/get/all', (req, res) => {
+router.get('/api/template/get_all', (req, res) => {
   TemplateDB.getAll((err, templates) => {
     if (!handleErr(err, res)) return;
-    console.log(templates);
     res.send({
-      documents: templates
+      templates: templates
     });
   });
 });
 
 // Get one document, returns all file objects.
 // Expects a 'name' parameter in body.
-router.post('/api/latex/get/one', (req, res) => {
+router.post('/api/template/get', (req, res) => {
   let name = req.body.name;
   if (name) {
     fs.readdir(Config.DATA_PATH + name, (err, files) => {
       if (!handleErr(err, res)) return;
 
-      let docFiles: File[] = [];
+      let docFiles: mFile[] = [];
       for (let f of files) {
         try {
           let m = mime.lookup(f);
@@ -135,28 +135,35 @@ router.post('/api/latex/get/one', (req, res) => {
 
 // Create a new document, returns an empty 'main.tex' file object.
 // Expects a 'name' parameter in body.
-router.post('/api/latex/create/one', (req, res) => {
+router.post('/api/template/create', (req, res) => {
   let name = req.body.name;
   if (name) {
+    // TODO: check db
+    if (fs.existsSync(Config.DATA_PATH + name)) {
+      res.status(500).send({
+        error: 'Template name already exists!'
+      });
+      return;
+    }
+
     async.parallel([
-      callback => {
-        TemplateDB.save({
-          name: name,
-          desc: req.body.desc,
-          active: true,
-        }, callback);
-      },
-      callback => {
-        fs.mkdir(Config.DATA_PATH + name, (err) => {
-          if (err) callback(err);
-          fs.writeFile(Config.DATA_PATH + name + '/' + 'main.tex', '', (err) => {
-            callback(err, [{name: 'main.tex', text: ''}]);
+        callback => { // save to db
+          TemplateDB.save({
+            name: name,
+            desc: req.body.desc,
+            active: true,
+          }, callback);
+        },
+        callback => { // create dir and main.tex
+          fs.mkdir(Config.DATA_PATH + name, (err) => {
+            if (err) callback(err);
+            fs.writeFile(Config.DATA_PATH + name + '/' + 'main.tex', '', (err) => {
+              callback(err, [{name: 'main.tex', text: ''}]);
+            });
           });
-        });
-      }],
-      (err, results) => {
+        }],
+      (err, results) => { // callback
         if (!handleErr(err, res)) return;
-        console.log(results);
         res.send({
           template: results[0],
           files: results[1],
@@ -169,11 +176,25 @@ router.post('/api/latex/create/one', (req, res) => {
 });
 
 // Delete one document, returns an empty response.
-// Expects a 'name' parameter in body.
-router.post('/api/latex/delete/one', (req, res) => {
-  let name = req.body.name;
-  if (name) {
-    removeDir(Config.DATA_PATH + name, res, () => res.sendStatus(200))
+// Expects the template to delete in body.
+router.post('/api/template/delete', (req, res) => {
+  let template: TemplateI = req.body.template;
+  if (template) {
+    async.parallel([
+        callback => { // delete from db
+          TemplateDB.remove(template, callback);
+        },
+        callback => { // delete from disk
+          fs.remove(Config.DATA_PATH + template.name, err => {
+            if (err) callback(err);
+            callback(undefined);
+          });
+        }
+      ],
+      (err, result) => { // callback
+        if (!handleErr(err, res)) return;
+        res.sendStatus(200);
+      });
   } else {
     console.log('Bad Request.');
     res.sendStatus(400);
@@ -181,37 +202,51 @@ router.post('/api/latex/delete/one', (req, res) => {
 });
 
 // Update a document, returns all updated file objects.
-// Expects a 'name' property and a 'files' property with an array of all file objects in body.
-router.post('/api/latex/update/one', (req, res) => {
-  let name: string = req.body.name;
-  let newFiles: File[] = req.body.files;
-  if (name && newFiles) {
-    if (!fs.existsSync(Config.DATA_PATH + name)) {
-      res.send({files: ''});
-      return;
-    }
-    fs.readdir(Config.DATA_PATH + name, (err, oldFiles) => {
-      if (!handleErr(err, res)) return;
-      try {
-        for (let of of oldFiles) {
-          if (!newFiles.find(f => f.name === of)) {
-            fs.unlinkSync(Config.DATA_PATH + name + '/' + of);
-            console.log('unlinked ' + of);
+// Expects a 'template' and a 'files' property with an array of all file objects in body.
+router.post('/api/template/update', (req, res) => {
+  let template: TemplateI = req.body.template;
+  let newFiles: mFile[] = req.body.files;
+  console.log('Update: ', req.body);
+  if (template && newFiles) {
+    async.parallel([
+        callback => { // update db
+          TemplateDB.change(template, callback);
+        },
+        callback => { // update files
+          if (!fs.existsSync(Config.DATA_PATH + template.name)) {
+            callback(`Could not find directory ${template.name}!`);
           }
+          fs.readdir(Config.DATA_PATH + template.name, (err, oldFiles) => {
+            if (err) callback(err);
+            try {
+              for (let of of oldFiles) {
+                if (!newFiles.find(f => f.name === of)) {
+                  fs.unlinkSync(Config.DATA_PATH + template.name + '/' + of);
+                  console.log('unlinked ' + of);
+                }
+              }
+              for (let nf of newFiles) {
+                let m = mime.lookup(nf.name);
+                if (m == 'image/png' || m == 'image/jpeg') {
+                  continue;
+                }
+                fs.writeFileSync(Config.DATA_PATH + template.name + '/' + nf.name, nf.text);
+              }
+            } catch (err) {
+              callback(err);
+            }
+            callback(undefined, newFiles);
+          })
         }
-        for (let nf of newFiles) {
-          let m = mime.lookup(nf.name);
-          if (m == 'image/png' || m == 'image/jpeg') {
-            continue;
-          }
-          fs.writeFileSync(Config.DATA_PATH + name + '/' + nf.name, nf.text);
-        }
-      } catch (err) {
-        console.error(err);
-        res.status(500).send(err);
-      }
-      res.send({files: newFiles});
-    })
+      ],
+      (err, result) => {
+        if (!handleErr(err, res)) return;
+        res.send({
+          template: result[0],
+          files: result[1],
+        })
+      });
+
   } else {
     console.log('Bad Request.');
     res.sendStatus(400);
@@ -224,15 +259,5 @@ function handleErr(err, res): boolean {
     res.status(500).send(err);
     return false;
   }
-  return true;
-}
-
-function removeDir(path: string, res, callback?): boolean {
-  fs.remove(path, err => {
-    if (!handleErr(err, res)) {
-      return false;
-    }
-    callback();
-  });
   return true;
 }
